rekajigger so the mime stuff is in an internal package
